+++ src/continuous_strategy.h
@@ -1,39 +1,39 @@
 //------------------------------------------------------------------------------
 //
-//  File:       categoricalStrategy.h
+//  File:       continuousStrategy.h
 //
-//  Description: strategy for categorical splits.
+//  Description: strategies for continuous splits.
 //
 //	Author: 	James Hickey
 //------------------------------------------------------------------------------
 
-#ifndef CATEGORICALSTRATEGY_H
-#define CATEGORICALSTRATEGY_H
+#ifndef CONTINUOUSSTRATEGY_H
+#define CONTINUOUSSTRATEGY_H
 
 //------------------------------
 // Includes
 //------------------------------
 #include "dataset.h"
 #include "node.h"
-#include "genericNodeStrategy.h"
+#include "generic_node_strategy.h"
 #include <Rcpp.h>
 
 
 //------------------------------
 // Class Definition
 //------------------------------
-class CategoricalStrategy:public GenericNodeStrategy
+class ContinuousStrategy:public GenericNodeStrategy
 {
 public:
 	//----------------------
 	// Public Constructors
 	//----------------------
-	CategoricalStrategy(CNode* node):nodeContext(node){};
+	ContinuousStrategy(CNode* node):nodeContext(node){};
 
 	//---------------------
 	// Public destructor
 	//---------------------
-	~CategoricalStrategy(){nodeContext=NULL;};
+	~ContinuousStrategy(){nodeContext=NULL;};
 
 	//---------------------
 	// Public Functions
@@ -61,10 +61,10 @@ public:
 		}
 	}
 	void Predict(const CDataset &data,
-				    unsigned long iRow,
-				    double &dFadj)
+			    unsigned long iRow,
+			    double &dFadj)
 	{
-		signed char schWhichNode = CategoricalStrategy::WhichNode(data,iRow);
+		signed char schWhichNode = ContinuousStrategy::WhichNode(data,iRow);
 		if(schWhichNode == -1)
 		{
 		  nodeContext->pLeftNode->Predict(data, iRow, dFadj);
@@ -123,23 +123,20 @@ public:
 	{
 		signed char ReturnValue = 0;
 		double dX = data.x_value(iObs, nodeContext->iSplitVar);
-
-    	if(!ISNA(dX))
-		{
-		  if(std::find(nodeContext->aiLeftCategory.begin(),
-			   nodeContext->aiLeftCategory.end(),
-			   (ULONG)dX) != nodeContext->aiLeftCategory.end())
+		 if(!ISNA(dX))
 			{
-				ReturnValue = -1;
+				if(dX < nodeContext->dSplitValue)
+				{
+					ReturnValue = -1;
+				}
+				else
+				{
+					ReturnValue = 1;
+				}
 			}
-			else
-			{
-				ReturnValue = 1;
-			}
-		}
-		// if missing value returns 0
+			// if missing value returns 0
 
-		return ReturnValue;
+			return ReturnValue;
 	}
 	void TransferTreeToRList
 	(
@@ -159,24 +156,11 @@ public:
 	)
 	{
 		int iThisNodeID = iNodeID;
-		unsigned long cCatSplits = vecSplitCodes.size();
-		unsigned long i = 0;
-		int cLevels = data.varclass(nodeContext->iSplitVar);
-		const std::size_t cLeftCategory = nodeContext->aiLeftCategory.size();
-
 		aiSplitVar[iThisNodeID] = nodeContext->iSplitVar;
-		adSplitPoint[iThisNodeID] = cCatSplits+cCatSplitsOld; // 0 based
+		adSplitPoint[iThisNodeID] = nodeContext->dSplitValue;
 		adErrorReduction[iThisNodeID] = nodeContext->dImprovement;
 		adWeight[iThisNodeID] = nodeContext->dTrainW;
 		adPred[iThisNodeID] = dShrinkage*nodeContext->dPrediction;
-
-		vecSplitCodes.push_back(VEC_CATEGORIES());
-
-		vecSplitCodes[cCatSplits].resize(cLevels,1);
-		for(i=0; i<cLeftCategory; i++)
-		  {
-			vecSplitCodes[cCatSplits][nodeContext->aiLeftCategory[i]] = -1;
-		  }
 
 		iNodeID++;
 		aiLeftNode[iThisNodeID] = iNodeID;
@@ -193,6 +177,7 @@ public:
 					 vecSplitCodes,
 					 cCatSplitsOld,
 					 dShrinkage);
+
 		aiRightNode[iThisNodeID] = iNodeID;
 		nodeContext->pRightNode->TransferTreeToRList(iNodeID,
 					  data,
@@ -207,7 +192,6 @@ public:
 					  vecSplitCodes,
 					  cCatSplitsOld,
 					  dShrinkage);
-
 		aiMissingNode[iThisNodeID] = iNodeID;
 		nodeContext->pMissingNode->TransferTreeToRList(iNodeID,
 						data,
@@ -227,4 +211,5 @@ public:
 private:
 	CNode* nodeContext;
 };
-#endif // CATEGORICALSTRATEGY_H
+#endif // CONTINUOUSSTRATEGY_H
+
