+++ src/gbm_functions.h
