+++ src/generic_node_strategy.h
