+++ src/node_parameters.h
